+++ src/index.ts
@@ -2,8 +2,14 @@ export { SkynetClient } from "./client";
 export { deriveChildSeed, genKeyPairAndSeed, genKeyPairFromSeed } from "./crypto";
 export { getRelativeFilePath, getRootDirectory } from "./utils/file";
 export { MAX_REVISION } from "./utils/number";
-export { parseSkylink, uriHandshakePrefix, uriHandshakeResolverPrefix, uriSkynetPrefix } from "./utils/skylink";
-export { defaultPortalUrl, defaultSkynetPortalUrl } from "./utils/url";
+export { parseSkylink } from "./utils/skylink";
+export {
+  defaultPortalUrl,
+  defaultSkynetPortalUrl,
+  uriHandshakePrefix,
+  uriHandshakeResolverPrefix,
+  uriSkynetPrefix,
+} from "./utils/url";
 
 // Export types.
