+++ src/utils/url.ts
@@ -4,6 +4,10 @@ import { trimSuffix } from "./string";
 
 export const defaultSkynetPortalUrl = "https://siasky.net";
 
+export const uriHandshakePrefix = "hns:";
+export const uriHandshakeResolverPrefix = "hnsres:";
+export const uriSkynetPrefix = "sia:";
+
 // TODO: This will be smarter. See
 // https://github.com/NebulousLabs/skynet-docs/issues/21.
 /**
